+++ src/app/data/about.js
@@ -1,22 +1,45 @@
 export const DATA_ABOUT = `
-PROCESSING ARCHITECTURE
 
-<br />
-For the first time Aarhus School of Architecture presents its traditional Graduation Show in a 100% digital format. The digital exhibition is a consequence of the lock down caused by COVID19, that with a flap of a butterfly’s wing, all of a sudden united and divided the world in to a new reality of fear, isolation and death, leaving stations, stadiums, schools and streets desolate, dramatically changing our behavior.
+CURATOR’S VOICE
 
 <br /><br />
-Under normal circumstances an exhibition on architecture consists of the physical encounter with drawings, models and materials, so, how can a digital exhibition replace that experience? The answer is: It cannot. But perhaps it can something else. In any case, every effort has been made to create an exhibition that embraces the spatial and the sensory, going from the usual physical support structures to a digital interface, developed by gamers and programmers.
+Processing Architecture 
 
 <br /><br />
-The question is, how will the temporarily imposed Social Distancing affect architecture in the future. Is it even possible to go from developing architecture that connects people, to architecture that disables human relations? And, what is the role of architecture in this challenging situation? It is too early to find the answers in the Exhibition, but you will find the answers to other urgent questions among this year´s projects.
+2020 radically changed our perception of space, from avoiding spaces we normally share with others to being repatriated, prevented from seeing colleagues, friends or family. From acting collectively on line, we have become collective on-liners. Meanwhile wars have continued, cities have been flooded, refugees have been refused entrance, and people have starved. Where do all these overwhelming issues leave architects and architecture? What perspectives and actions should be taken? And how do the economic and ecologic crisis affect our cities, culture, and eventually democracy?
 
 <br /><br />
-WDAcomprises a space of columns designed specifically for the Exhibition by the new Architects, where the school constitutes the foundation of the space, – but they are the pillars on which the architecture of tomorrow is built.
+Big questions, big challenges - and lots of work to be done. But also lots of relevant architectural matters to be discussed before initiating. Dialogues with the new architects about their projects should therefore have been filmed and available here on the Online Exhibition Site, but like so many other things filming was cancelled due to COVID 19. That means you can ‘only’ scroll through the projects  and experience how the new architects Process Architecture, eventually download the full thesises, and hopefully be inspired to joint venture with a new employee.
 
 <br /><br />
-WELCOME > THE EXHIBITION IS (ALWAYS) OPEN!
+So how do architects process architecture in times of crisis and how can architecture be part of the solution? Well, fundamentally architects have the ability to embed thought or emotion in space and buildings, they are aware that we are continually shaped by our surroundings, and that architecture can shift corporate cultures, lift spirits and impact the wellbeing of people. That´s why architects, urban landscapers or city planners are part of the solution. 
 
 <br /><br />
-Karen Kjaergaard, WDA2020
+This year’s projects reveal a sincere interest in contributing to a better world, were social relations rather than objectified architecture matter. The new architects are occupied with nature and the creation of cultural landscapes, but at the same time they master digital tooling. They have empathy and a willingness to make a change. 
+
+<br /><br />
+The New Architects are ready to Process Architecture. 
+
+<br /><br />
+Welcome to the Online Exhibition Site<br />
+Karen Kjaergaard, Curator<br />
+January 2021
+
+
+<br /><br /><br /><br />
+
+CREDITS
+
+<br /><br />
+Curatorial Concept: Karen Kjaergaard, curator & Head of Exhibitions, Aarhus School of Architecture
+
+<br /><br />
+Graphics: Casper Riis / <a href="https://www.walk.agency/" style="color:#fff; text-decoration:underline;">Walk Agency</a>
+
+<br /><br />
+Programming: Jørgen Skogmo / <a href="https://dearstudio.dk/" style="color:#fff; text-decoration:underline;">dearstudio</a>
+
+<br /><br />
+Lars Steen Møller / <a href="https://www.apparat.dk/" style="color:#fff; text-decoration:underline;">Apparat</a>
 
 `