+++ src/app/data/students-bak.js
@@ -1,0 +1,461 @@
+
+export const DATA_ORIGIN   = "Afgang Efteråret 2020_inkl forsinkede.xlsx"
+export const DATA_PARSEDATE = "Tue Dec 08 2020"
+export const DATA_STUDENTS = [
+  {
+    "id": "4008524",
+    "studio": "1A",
+    "name": "Tim Greentree",
+    "theme": "1",
+    "title": "(Perma)culture",
+    "stub": "tim-greentree"
+  },
+  {
+    "id": "4000500",
+    "studio": "1A",
+    "name": "Nikolai Thyregaard Nielsen",
+    "theme": "5",
+    "title": "Total Landscape - Case Study Herning",
+    "stub": "nikolai-thyregaard-nielsen"
+  },
+  {
+    "id": "4003467",
+    "studio": "1A",
+    "name": "Sofie Dalsgaard Jensen",
+    "theme": "3",
+    "title": "Når borgerne driver værket / En selvbygget bydel i Århus",
+    "stub": "sofie-dalsgaard-jensen"
+  },
+  {
+    "id": "4004320",
+    "studio": "1A",
+    "name": "Christine Berg Steffensen",
+    "theme": "5",
+    "title": "Det stille LANDSKAB - Jammerbugten",
+    "stub": "christine-berg-steffensen"
+  },
+  {
+    "id": "4001986",
+    "studio": "1A",
+    "name": "Maria Lunde Ørris",
+    "theme": "5",
+    "title": "Vildrenens forsvundne territorier",
+    "stub": "maria-lunde-oerris"
+  },
+  {
+    "id": "4008528",
+    "studio": "1A",
+    "name": "Eleanor Margot Gibson",
+    "theme": "5",
+    "title": "Reimagining the Company Town: A model for future landscapes of coexistence",
+    "stub": "eleanor-margot-gibson"
+  },
+  {
+    "id": "4003893",
+    "studio": "1A",
+    "name": "Sebastian Remme Yeoman",
+    "theme": "3",
+    "title": "EN BYDEL FORANKRET I HISTORIEN - Transformation af Aalborg Universitetshospital Nord",
+    "stub": "sebastian-remme-yeoman"
+  },
+  {
+    "id": "4007496",
+    "studio": "1A",
+    "name": "Huiru Huang",
+    "theme": "5",
+    "title": "Responsible Mining Landscape in Greenland for Sustainable Future",
+    "stub": "huiru-huang"
+  },
+  {
+    "id": "4002812",
+    "studio": "1A",
+    "name": "Simone Stellô Stelsø Lauridsen",
+    "theme": "5",
+    "title": "LAND & VAND - Landskabet på MOLS",
+    "stub": "simone-stello-stelsoe-lauridsen"
+  },
+  {
+    "id": "4003455",
+    "studio": "1B",
+    "name": "Vibeke Vildbrad Kristiansen",
+    "theme": "6",
+    "title": "Restaurering af Klingebækgård",
+    "stub": "vibeke-vildbrad-kristiansen"
+  },
+  {
+    "id": "4003574",
+    "studio": "1B",
+    "name": "Solveig Vangen Paulsen",
+    "theme": "6",
+    "title": "Deichmanske Bibliotek",
+    "stub": "solveig-vangen-paulsen"
+  },
+  {
+    "id": "4003590",
+    "studio": "1B",
+    "name": "Tora Hay Walseng",
+    "theme": "6",
+    "title": "Deichmanske Bibliotek",
+    "stub": "tora-hay-walseng"
+  },
+  {
+    "id": "4004125",
+    "studio": "1C",
+    "name": "Nora Regine Sund Dyrvik",
+    "theme": "2",
+    "title": "Fra museum til musikk- og kulturhus. En radikal transformasjon av Munchmuseet",
+    "stub": "nora-regine-sund-dyrvik"
+  },
+  {
+    "id": "4003780",
+    "studio": "1C",
+    "name": "Patrick Bunke",
+    "theme": "3",
+    "title": "Immervad - Et Klimatorium i Ådalen",
+    "stub": "patrick-bunke"
+  },
+  {
+    "id": "4003939",
+    "studio": "1C",
+    "name": "Thomas Maagaard",
+    "theme": "3",
+    "title": "Immervad - Et Klimatorium i Ådalen",
+    "stub": "thomas-maagaard"
+  },
+  {
+    "id": "4003444",
+    "studio": "1C",
+    "name": "Simon Rode Gregersen",
+    "theme": "2",
+    "title": "Krigens lag – transformation af Field Station Berlin",
+    "stub": "simon-rode-gregersen"
+  },
+  {
+    "id": "4002854",
+    "studio": "1C",
+    "name": "Tonje Benedicte Madsen",
+    "theme": "2",
+    "title": "Akropolis-høyden – En radikal transformasjon av et kulturelt landemerke i Stavanger",
+    "stub": "tonje-benedicte-madsen"
+  },
+  {
+    "id": "4003564",
+    "studio": "1C",
+    "name": "Silja Nordvåg",
+    "theme": "2",
+    "title": "Alfheim Satellitt Universitet - en radikal transformasjon av en senmodernistisk svømmehall",
+    "stub": "silja-nordvaag"
+  },
+  {
+    "id": "4004177",
+    "studio": "1C",
+    "name": "Kathrine Grundahl Hansen",
+    "theme": "1",
+    "title": "ET MARITIMT CENTER - En transformation af siloerne ved Lillebælt",
+    "stub": "kathrine-grundahl-hansen"
+  },
+  {
+    "id": "4003946",
+    "studio": "1C",
+    "name": "Jacob Haugaard Mikkelsen",
+    "theme": "3",
+    "title": "Ankomsten til Aarhus fra Øst - Omdannelse af silokompleks til terminal for skibsturisme",
+    "stub": "jacob-haugaard-mikkelsen"
+  },
+  {
+    "id": "4002766",
+    "studio": "2A",
+    "name": "Julie Ærthøj Velling",
+    "theme": "4",
+    "title": "A Matter of Proportions",
+    "stub": "julie-aerthoej-velling"
+  },
+  {
+    "id": "4000127",
+    "studio": "2A",
+    "name": "Anne Søby Nielsen",
+    "theme": "5",
+    "title": "Architecture for Contested Sites: the Case of Grindsted",
+    "stub": "anne-soeby-nielsen"
+  },
+  {
+    "id": "4003774",
+    "studio": "2A",
+    "name": "Lucas Philip Edwards",
+    "theme": "2",
+    "title": "Song to Song — an observatory for forestry",
+    "stub": "lucas-philip-edwards"
+  },
+  {
+    "id": "4003790",
+    "studio": "2A",
+    "name": "Malene Thornemann Johansen",
+    "theme": "2",
+    "title": "Formet af vand - Et center for bevægelse",
+    "stub": "malene-thornemann-johansen"
+  },
+  {
+    "id": "4004348",
+    "studio": "2A",
+    "name": "Mathilde Ballegaard Lehrmann",
+    "theme": "1",
+    "title": "In between life and death - Spaces of transition",
+    "stub": "mathilde-ballegaard-lehrmann"
+  },
+  {
+    "id": "4004084",
+    "studio": "2A",
+    "name": "Julie Høj Holk",
+    "theme": "1",
+    "title": "Motivation til interaktion - et ressourcestærkt fællesskab",
+    "stub": "julie-hoej-holk"
+  },
+  {
+    "id": "4003465",
+    "studio": "2A",
+    "name": "Alexander Hejgaard Fagerholt",
+    "theme": "3",
+    "title": "In between typologies - The future challenges at Refshaleøen",
+    "stub": "alexander-hejgaard-fagerholt"
+  },
+  {
+    "id": "4003484",
+    "studio": "2A",
+    "name": "Pernille Dambo",
+    "theme": "4",
+    "title": "From Linear to Circular - A rethinking of the waste station",
+    "stub": "pernille-dambo"
+  },
+  {
+    "id": "4006503",
+    "studio": "2A",
+    "name": "Rosa Prichard",
+    "theme": "5",
+    "title": "Decentering The Visitor: A Critical Reframing of Bornholm's Extracted Landscapes",
+    "stub": "rosa-prichard"
+  },
+  {
+    "id": "4003709",
+    "studio": "2B",
+    "name": "Elias Johan Hamann",
+    "theme": "7",
+    "title": "RSDO-1-21 / ENVISIONING THE RECONSTRUCTION OF SPACE DEBRIS IN ORBIT TO A SPACE HABITAT",
+    "stub": "elias-johan-hamann"
+  },
+  {
+    "id": "4002805",
+    "studio": "2B",
+    "name": "Jesper Asferg Scheel",
+    "theme": "2",
+    "title": "‘ROUND AND ROUND WE GO’ - The velodrome imagined as an homage to the legacy, sounds and emotions of cycling.",
+    "stub": "jesper-asferg-scheel"
+  },
+  {
+    "id": "4004000",
+    "studio": "2B",
+    "name": "Jens Toft Madsen",
+    "theme": "2",
+    "title": "New Arrivals - Establishing a connection between sea & city",
+    "stub": "jens-toft-madsen"
+  },
+  {
+    "id": "4004158",
+    "studio": "2B",
+    "name": "Carl Jørgen Steen Holst",
+    "theme": "4",
+    "title": "Disappearing Snow - The Norwegian Ski Culture in a Changing Climate",
+    "stub": "carl-joergen-steen-holst"
+  },
+  {
+    "id": "4003805",
+    "studio": "2B",
+    "name": "Kristoffer Holmgaard Gade",
+    "theme": "7",
+    "title": "AN UNWELCOME GUEST - A PARASITIC HABITAT FOR A NUCLEAR REACTOR",
+    "stub": "kristoffer-holmgaard-gade"
+  },
+  {
+    "id": "4000080",
+    "studio": "2C",
+    "name": "Chris Jongberg Johansen",
+    "theme": "1",
+    "title": "Ask cohousing- A part of the village Ask bofællesskab",
+    "stub": "chris-jongberg-johansen"
+  },
+  {
+    "id": "4003408",
+    "studio": "2C",
+    "name": "Thorlak Solberg",
+    "theme": "",
+    "title": "Tættere på naturen: Højskolen som mødested og mulighedsrum",
+    "stub": "thorlak-solberg"
+  },
+  {
+    "id": "4003696",
+    "studio": "2C",
+    "name": "Sarah Knudsen Ghobadi-Salehbeglou",
+    "theme": "1",
+    "title": "Den Nære Arbejdsplads: Et Bokoncept til fremme af Hjemmearbejde og Lokalmiljø",
+    "stub": "sarah-knudsen-ghobadi-salehbeglou"
+  },
+  {
+    "id": "4003277",
+    "studio": "2C",
+    "name": "Jesper Andreas Christiansen",
+    "theme": "",
+    "title": "Dumb Suburbia? Recalibrating the suburban settlement of Skjoldhøjparken",
+    "stub": "jesper-andreas-christiansen"
+  },
+  {
+    "id": "4003432",
+    "studio": "2C",
+    "name": "Jakob Vad",
+    "theme": "",
+    "title": "Dumb Suburbia? Recalibrating the suburban settlement of Skjoldhøjparken",
+    "stub": "jakob-vad"
+  },
+  {
+    "id": "4003739",
+    "studio": "2C",
+    "name": "Morten Aggerholm Andersen",
+    "theme": "1",
+    "title": "EN MENINGSGIVENDE FUSIONERING AF TO KOMUNALE INSTITUTION",
+    "stub": "morten-aggerholm-andersen"
+  },
+  {
+    "id": "4003759",
+    "studio": "2C",
+    "name": "Simon Kobberup Smedsmo",
+    "theme": "2",
+    "title": "Sport- og naturfaciliteter i søområde i Holstebro",
+    "stub": "simon-kobberup-smedsmo"
+  },
+  {
+    "id": "4003461",
+    "studio": "2C",
+    "name": "Khadar Yusuf Awil",
+    "theme": "2",
+    "title": "Livet mellem reolerne: et lokalbibliotek i Aalborg Vestby",
+    "stub": "khadar-yusuf-awil"
+  },
+  {
+    "id": "4003812",
+    "studio": "2C",
+    "name": "Tanja Tofte Thomsen",
+    "theme": "2",
+    "title": "Et naturvandrehjem på Røsnæs - med plads til fordybelse",
+    "stub": "tanja-tofte-thomsen"
+  },
+  {
+    "id": "4003817",
+    "studio": "2C",
+    "name": "Helena Feld",
+    "theme": "1",
+    "title": "Veteranernes Hjem - naturens heling af krigens sår",
+    "stub": "helena-feld"
+  },
+  {
+    "id": "4003493",
+    "studio": "2C",
+    "name": "Clara Holm Lange",
+    "theme": "3",
+    "title": "Nyholm i Forandring - Tradition, kultur og uddannelse",
+    "stub": "clara-holm-lange"
+  },
+  {
+    "id": "4003872",
+    "studio": "2C",
+    "name": "Tanja Haugstrup Hansen",
+    "theme": "",
+    "title": "3 generationer under samme tag - på kanten af by og land",
+    "stub": "tanja-haugstrup-hansen"
+  },
+  {
+    "id": "4003889",
+    "studio": "2C",
+    "name": "Daniel Nielsen",
+    "theme": "1",
+    "title": "House, two Homes",
+    "stub": "daniel-nielsen"
+  },
+  {
+    "id": "4002808",
+    "studio": "2C",
+    "name": "Vilde Livsdatter Sønderland",
+    "theme": "1",
+    "title": "Food as a social catalyst in an urban environment",
+    "stub": "vilde-livsdatter-soenderland"
+  },
+  {
+    "id": "4004118",
+    "studio": "3",
+    "name": "Karl Henrik Kobbeltvedt",
+    "theme": "",
+    "title": "Knubben - The rehabilitation of an old essential",
+    "stub": "karl-henrik-kobbeltvedt"
+  },
+  {
+    "id": "4003416",
+    "studio": "3",
+    "name": "Malene Jørs Nielsen",
+    "theme": "4",
+    "title": "A Tidal Encounter on Mandø",
+    "stub": "malene-joers-nielsen"
+  },
+  {
+    "id": "4003711",
+    "studio": "3",
+    "name": "Marlene Abild Hindsted",
+    "theme": "4",
+    "title": "A Production from the Seabed",
+    "stub": "marlene-abild-hindsted"
+  },
+  {
+    "id": "4004042",
+    "studio": "3",
+    "name": "Charlotte Lyberth Jensen",
+    "theme": "",
+    "title": "friRUM - a new join-in centre for socially vulnerable in Aarhus",
+    "stub": "charlotte-lyberth-jensen"
+  },
+  {
+    "id": "4003445",
+    "studio": "3",
+    "name": "Michelle Skov Justesen",
+    "theme": "3",
+    "title": "Rebirth and transformation of Randers Harbour",
+    "stub": "michelle-skov-justesen"
+  },
+  {
+    "id": "4001983",
+    "studio": "3",
+    "name": "Stine Nørgaard Skovbjerg",
+    "theme": "",
+    "title": "The New Old Way - A School for Self-Sustainable Farming",
+    "stub": "stine-noergaard-skovbjerg"
+  },
+  {
+    "id": "4003441",
+    "studio": "2A",
+    "name": "Michelle Rosenvinge",
+    "theme": "1",
+    "title": "Perception in the Absence of Sight : Rehabilitation for the Visually Impaired",
+    "stub": "michelle-rosenvinge"
+  },
+  {
+    "id": "4002814",
+    "studio": "1B",
+    "name": "Lisa Christine Sippli",
+    "theme": "",
+    "title": "Zeche Gneisenau - en revitalisering af det gamle kulmineanlæg i Dortmund Derne",
+    "stub": "lisa-christine-sippli"
+  },
+  {
+    "id": "4003947",
+    "studio": "2C",
+    "name": "Mischa Josefine Stæhr",
+    "theme": "",
+    "title": "",
+    "stub": "mischa-josefine-staehr"
+  }
+]