+++ src/app/settings.js
@@ -1,5 +1,8 @@
 export const settings = {
 	document_title: 'PROCESSING ARCHITECTURE - AAA21',
-	title: 	`//&nbsp;PROCESSING ARCHITECTURE`,
+	// title: 	`//&nbsp;PROCESSING ARCHITECTURE`,
+	// title: 	`/** &nbsp;*&nbsp;2020/2021 */ PROCESSING ARCHITECTURE`,
+	title: 	`//&nbsp;2020/2021 //&nbsp;PROCESSING //&nbsp;ARCHITECTURE`,
+	// title: 	`/* *&nbsp;2020/2021 *&nbsp;PROCESSING *&nbsp;ARCHITECTURE **/`,
 
 }